pull out cat class
